# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: AgentOps Desktop HTTP API
  description: |
    HTTP REST API for AgentOps Desktop — manage AI agents, goals, tasks, schedules, squads, cost budgets, adapters, and skills.

    **Base URL:** `http://localhost:3967`
    **Authentication:** Bearer token via `POST /auth/login`. Include as `Authorization: Bearer <token>` on all `/api/*` routes.
    **Default port:** 3967 (configurable via `PORT` env var)
  version: "0.2.0"
  contact:
    name: AgentOps Team
  license:
    name: UNLICENSED
servers:
  - url: "http://localhost:3967"
    description: Local HTTP API server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login
  schemas:
...
    Goal:
      type: object
      properties:
        id:
          type: string
          example: "goal-1"
        title:
          type: string
          example: Ship MVP
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, archived]
          example: active
        taskIds:
          type: array
          items:
            type: string
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
paths:
...
  /api/goals:
    get:
      summary: List goals
      operationId: goalsList
      tags: [Goals]
      security:
        - bearerAuth: []
      parameters:
        - { name: offset, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: status, in: query, schema: { type: string, enum: [active, completed, archived] } }
        - { name: sortBy, in: query, schema: { type: string, enum: [title, status, createdAt, updatedAt] } }
        - { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "200":
          description: Goal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Goal" }
    post:
      summary: Create goal
      operationId: goalsCreate
      tags: [Goals]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 500 }
                description: { type: string, maxLength: 5000 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: "#/components/schemas/Goal" }
  /api/goals/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get goal
      operationId: goalsGet
      tags: [Goals]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Goal
        "404":
          description: Not found
    patch:
      summary: Update goal
      operationId: goalsUpdate
      tags: [Goals]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                status: { type: string, enum: [active, completed, archived] }
      responses:
        "200":
          description: Updated
        "404":
          description: Not found
    delete:
      summary: Delete goal
      operationId: goalsDelete
      tags: [Goals]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found

  # ── Tasks ──
...
  /api/cost/budgets:
    get:
      summary: List budgets
      operationId: costBudgetsList
      tags: [Cost]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Budget list
    post:
      summary: Create budget
      operationId: costBudgetsCreate
      tags: [Cost]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId, monthlyLimit]
              properties:
                agentId: { type: string }
                monthlyLimit: { type: number }
                currency: { type: string }
                warnPct: { type: number }
                pausePct: { type: number }
                stopPct: { type: number }
      responses:
        "201":
          description: Created
        "409":
          description: Budget already exists for agent
  /api/cost/budgets/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get budget
      operationId: costBudgetsGet
      tags: [Cost]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Budget
        "404":
          description: Not found
    patch:
      summary: Update budget
      operationId: costBudgetsUpdate
      tags: [Cost]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                monthlyLimit: { type: number }
                currency: { type: string }
                warnPct: { type: number }
                pausePct: { type: number }
                stopPct: { type: number }
      responses:
        "200":
          description: Updated
        "404":
          description: Not found
    delete:
      summary: Delete budget
      operationId: costBudgetsDelete
      tags: [Cost]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found
  /api/cost/budgets/{id}/reset:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      summary: Reset monthly budget
      operationId: costBudgetsReset
      tags: [Cost]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Reset
        "404":
          description: Not found
  /api/cost/budgets/agent/{agentId}:
    parameters:
      - { name: agentId, in: path, required: true, schema: { type: string } }
    get:
      summary: Get budget by agent
      operationId: costBudgetsByAgent
      tags: [Cost]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Budget
        "404":
          description: Not found
  /api/cost/usage:
    get:
      summary: List usage records
      operationId: costUsageList
      tags: [Cost]
      security:
        - bearerAuth: []
      parameters:
        - { name: agentId, in: query, schema: { type: string } }
        - { name: taskId, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        "200":
          description: Usage records
        "422":
          description: agentId or taskId required
    post:
      summary: Log usage
      operationId: costUsageLog
      tags: [Cost]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId]
              properties:
                agentId: { type: string }
                taskId: { type: string }
                inputTokens: { type: number }
                outputTokens: { type: number }
                costUsd: { type: number }
                model: { type: string }
                provider: { type: string }
      responses:
        "201":
          description: Usage logged
  /api/cost/reports:
    get:
      summary: Cost reports
      operationId: costReports
      tags: [Cost]
      security:
        - bearerAuth: []
      parameters:
        - { name: agentId, in: query, schema: { type: string } }
        - { name: goalId, in: query, schema: { type: string } }
        - { name: since, in: query, schema: { type: string } }
        - { name: until, in: query, schema: { type: string } }
      responses:
        "200":
          description: Cost report
  /api/cost/reports/agent/{agentId}/period:
    parameters:
      - { name: agentId, in: path, required: true, schema: { type: string } }
    get:
      summary: Spend by period for agent
      operationId: costReportsAgentPeriod
      tags: [Cost]
      security:
        - bearerAuth: []
      parameters:
        - { name: since, in: query, schema: { type: string } }
        - { name: until, in: query, schema: { type: string } }
      responses:
        "200":
          description: Spend by period

  # ── Adapters ──
...